major refactor, better params, several bug fixes
I refactored the code because it was hard to debug
and hard to understand what was going wrong.

Changes include:

- better argument input and documentation
- args now use flags, positionals were wonky and error prone
- error checking on inputs
- added -v, --verbose flag
- added logging module w/ -v
- better error handling
- added --ttl flag
- token as an arg, not env var
- bugfix: fmt.Sprintf to convert account int
- bugfix: do not require setting all record vars (regions bug)

---

Why token as an arg?

env vars leak, they are bad for common usage. better to:

    dnslink-dnsimple -t $(cat .dnsimple-token)

Yes, infra deployments tend to use env vars. it's easy to
turn an env var into an arg for such use cases:

    dnslink-dnsimple -t $(echo DNSIMPLE_TOKEN)

But it is hard to avoid the leakage in other settings.
-